move uuid to extensions dir
